xadmin-server-fastapi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

sqlalchemy==2.0.23
alembic==1.13.1
pymysql==1.1.0
aiosqlite==0.19.0

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2
cryptography==45.0.6
pycryptodome==3.19.0

redis==5.0.1
aioredis==2.0.1

pillow==10.1.0
aiofiles==23.2.1
openpyxl==3.1.2

httpx==0.25.2
requests==2.31.0

python-dotenv==1.0.0
click==8.1.7
email-validator==2.1.0
phonenumbers==8.13.25
user-agents==2.2.0
psutil==5.9.6

captcha==0.5.0
qrcode==7.4.2

websockets==12.0